+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.176.3] - 2024-04-18
+### Changed
+- Updated KAMU_WEB_UI image to latest release 0.18.1
+
 ## [0.176.2] - 2024-04-16
 ### Fixed
 - Fix the instant run of `ExecuteTransform` flow after the successful finish of `HardCompacting` flow